% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calNDSI.R
\name{calNDSI}
\alias{calNDSI}
\title{Calculates Normalized Difference Snow Index (NDSI)}
\usage{
calNDSI(R2, R5)
}
\arguments{
\item{R2}{The reflectance from channel 2}

\item{R5}{The reflectance from channel 5}
}
\description{
Calculates Normalized Difference Snow Index (NDSI)
}
