% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calNDVI.R
\name{calNDVI}
\alias{calNDVI}
\title{Calculates Normalized Difference Vegetation Index (NDVI)}
\usage{
calNDVI(R2, R3)
}
\arguments{
\item{R2}{The reflectance from channel 2}

\item{R3}{The reflectance from channel 3}
}
\description{
Calculates Normalized Difference Vegetation Index (NDVI)
}
