% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createBayesModel_DB.R
\name{createBayesModel.DB}
\alias{createBayesModel.DB}
\title{Create a Bayes Model for a deciduous broadleaf site}
\usage{
createBayesModel.DB(
  dataSource,
  siteName = "",
  URL = "",
  niter = 1e+05,
  startDate,
  endDate,
  lat,
  long,
  TZ = 5,
  PFT = "DB",
  maxValue = FALSE,
  seasonOrder = "FS"
)
}
\arguments{
\item{dataSource}{data source (GOES.NDVI, MODIS.NDVI, PC.GCC)}

\item{siteName}{Site Name}

\item{URL}{PhenoCam network URL}

\item{niter}{the maximum number of iterations you want to give the model to converge within}

\item{startDate}{The date to start the model}

\item{endDate}{The date to end the model}

\item{lat}{latitude of site in degrees}

\item{long}{longitude of site in degrees}

\item{seasonOrder}{"FS" or "SF"}
}
\description{
Create a Bayes Model for a deciduous broadleaf site
}
