% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createModel_DB_Avg.R
\name{createBayesModel.DB_Avg}
\alias{createBayesModel.DB_Avg}
\title{Create a Bayes Model for a deciduous broadleaf site using the window average GOES values}
\usage{
createBayesModel.DB_Avg(dataSource = "GOES.NDVI", siteName = "", URL = "",
  niter = 1e+05, startDay, endDay, lat, long, TZ = 5)
}
\arguments{
\item{dataSource}{data source (GOES.NDVI, MODIS.NDVI, PC.GCC)}

\item{siteName}{Site Name}

\item{URL}{PhenoCam network URL}

\item{niter}{the maximum number of iterations you want to give the model to converge within}

\item{startDay}{the day of year since 2017-01-01 to start the model}

\item{endDay}{the day of year since 2017-01-01 to end the model}

\item{lat}{latitude of site in degrees}

\item{long}{longitude of site in degrees}
}
\description{
Create a Bayes Model for a deciduous broadleaf site using the window average GOES values
}
