% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createOverallModel_IncludeDiurnal.R
\name{createBayesModel.DB_Overall}
\alias{createBayesModel.DB_Overall}
\title{Create a Bayes Model for a deciduous broadleaf site based on the diurnal-fit NDVI data}
\usage{
createBayesModel.DB_Overall(data, niter = 1e+05)
}
\arguments{
\item{niter}{the maximum number of iterations you want to give the model to converge within}

\item{dataSource}{data source (GOES.NDVI, MODIS.NDVI, PC.GCC)}

\item{siteName}{Site Name}

\item{URL}{PhenoCam network URL}

\item{lat}{latitude of site in degrees}

\item{long}{longitude of site in degrees}

\item{startDay}{the day of year since 2017-01-01 to start the model}

\item{endDay}{the day of year since 2017-01-01 to end the model}
}
\description{
Create a Bayes Model for a deciduous broadleaf site based on the diurnal-fit NDVI data
}
