% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createModel_Diurnal.R
\name{createBayesModel.Diurnal}
\alias{createBayesModel.Diurnal}
\title{Create a diurnal model for a deciduous broadleaf site}
\usage{
createBayesModel.Diurnal(siteName, data)
}
\arguments{
\item{siteName}{Site Name}

\item{data}{Data object}
}
\description{
Create a diurnal model for a deciduous broadleaf site
}
