% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createModel_EN.R
\name{createBayesModel.EN}
\alias{createBayesModel.EN}
\title{Create a Bayes Model for an evergreen needleleaf site}
\usage{
createBayesModel.EN(dataSource, siteName = "", URL = "")
}
\arguments{
\item{dataSource}{data source (GOES.NDVI, MODIS.NDVI, PC.GCC)}

\item{siteName}{Site Name}

\item{URL}{PhenoCam network URL}
}
\description{
Create a Bayes Model for an evergreen needleleaf site
}
