% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createModel_SH.R
\name{createBayesModel.SH}
\alias{createBayesModel.SH}
\title{Create a Bayes Model for a shrubland site}
\usage{
createBayesModel.SH(dataSource, siteName = "", URL = "", startDay, endDay,
  lat, long, TZ = 5)
}
\arguments{
\item{dataSource}{data source (GOES.NDVI, MODIS.NDVI, PC.GCC)}

\item{siteName}{Site Name}

\item{URL}{PhenoCam network URL}

\item{startDay}{the day of year since 2017-01-01 to start the model}

\item{endDay}{the day of year since 2017-01-01 to end the model}

\item{lat}{latitude of site in degrees}

\item{long}{longitude of site in degrees}

\item{niter}{the maximum number of iterations you want to give the model to converge within}
}
\description{
Create a Bayes Model for a shrubland site
}
