% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCI.R
\name{createCI}
\alias{createCI}
\title{Create the credible interval from MCMC output matrix}
\usage{
createCI(PFT, var.mat, xseq, doRescale = TRUE, seasonOrder = "FS")
}
\arguments{
\item{PFT}{The plant functional type (DB or SH)}

\item{var.mat}{The MCMC output as a matrix}

\item{xseq}{The sequence of x values you want to calculate the CI for}

\item{doRescale}{Boolean for if you want to rescale the CI to be between 0 and 1 (default=True)}

\item{seasonOrder}{the order of seasons FS or SF}
}
\description{
Create the credible interval from MCMC output matrix
}
