% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDBModel_ALL.R
\name{createDBModel_ALL}
\alias{createDBModel_ALL}
\title{Create a Bayes Model for a deciduous broadleaf site using all data sources}
\usage{
createDBModel_ALL(siteName = "", URL = "", niter = 1e+05, startDay,
  endDay, lat, long, TZ = 5, seasonOrder = "FS", dataDirectory)
}
\arguments{
\item{siteName}{Site Name}

\item{URL}{PhenoCam network URL}

\item{niter}{the maximum number of iterations you want to give the model to converge within}

\item{startDay}{the day of year since 2017-01-01 to start the model}

\item{endDay}{the day of year since 2017-01-01 to end the model}

\item{lat}{latitude of site in degrees}

\item{long}{longitude of site in degrees}

\item{seasonOrder}{The order of seasons (default is "FS" for fall and then spring)}

\item{dataDirectory}{The data directory}
}
\description{
Create a Bayes Model for a deciduous broadleaf site using all data sources
}
