% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createInits.R
\name{createInits}
\alias{createInits}
\title{Function to create initial conditions}
\usage{
createInits(data, PFT)
}
\arguments{
\item{data}{The data in the form of a list with items x and y}

\item{PFT}{Plant functional type ("SH" or "DB")}
}
\description{
Function to create initial conditions
}
