% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createModel_DB.R
\name{createModel_DB}
\alias{createModel_DB}
\title{Create a Bayes Model for a deciduous broadleaf site}
\usage{
createModel_DB(data, dataSource, seasonOrder)
}
\arguments{
\item{data}{Data in the form of list that has data$y, data$x, and data$n defined}

\item{dataSource}{data source (GOES.NDVI, MODIS.NDVI, PC.GCC, MODIS.EVI)}

\item{seasonOrder}{The season order ("SF" or "FS")}
}
\description{
Create a Bayes Model for a deciduous broadleaf site
}
