% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createModelDB_spring.R,
%   R/createModelDB_springFall.R
\name{createModelDB_springFall}
\alias{createModelDB_springFall}
\alias{createModelDB_springFall}
\title{Create a spring Bayes Model for a deciduous broadleaf site}
\usage{
createModelDB_springFall(data, dataSource)

createModelDB_springFall(data, dataSource)
}
\arguments{
\item{data}{Data in the form of list that has data$y, data$x, and data$n defined}

\item{dataSource}{data source (GOES.NDVI, MODIS.NDVI, PC.GCC, MODIS.EVI)}

\item{data}{Data in the form of list that has data$y, data$x, and data$n defined}

\item{dataSource}{data source (GOES.NDVI, MODIS.NDVI, PC.GCC, MODIS.EVI)}
}
\description{
Create a spring Bayes Model for a deciduous broadleaf site

Create a spring Bayes Model for a deciduous broadleaf site
}
