% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createNDSI.R
\name{createNDSI}
\alias{createNDSI}
\title{Function to create a file with NDSI values for a specific day and time}
\usage{
createNDSI(day.time)
}
\arguments{
\item{day.time}{The day and time value in the format year,DOY,time (e.g. "20170151200" or "20171821658")}
}
\description{
Function to create a file with NDSI values for a specific day and time
}
