% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createOutputSummary.R
\name{createOutputSummary}
\alias{createOutputSummary}
\title{Creates an output summary (not finalized)}
\usage{
createOutputSummary(siteFileName, iseq, startDay, endDay)
}
\arguments{
\item{siteFileName}{The desired output file name}

\item{startDay}{The start day}

\item{endDay}{The end day}

\item{siteName}{iseq The row numbers of the sites in the data file to include in the summary}
}
\description{
Creates an output summary (not finalized)
}
