% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitEquations.R
\name{deciduousYvals}
\alias{deciduousYvals}
\title{Compute the double logistic curve (autumn then spring) for phenology values}
\usage{
deciduousYvals(TranF, bF, TranS, bS, c, d, k, xseq, seasonOrder = "FS")
}
\arguments{
\item{TranF}{DOY of autumn/fall transition date}

\item{bF}{autumn/fall b parameter (from Zhang et al., 2003)}

\item{TranS}{DOY of spring transition date}

\item{bS}{spring b parameter (from Zhang et al., 2003)}

\item{c}{c parameter (from Zhang et al., 2003)}

\item{d}{d parameter (from Zhang et al., 2003)}

\item{k}{k (changepoint) parameter}

\item{xseq}{the sequence of x values to compute the double logistic curve for}

\item{seasonOrder}{the order of seasons FS or SF}
}
\description{
Compute the double logistic curve (autumn then spring) for phenology values
}
