% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitEquations.R
\name{gaussianCurve}
\alias{gaussianCurve}
\title{Compute the gaussian curve values}
\usage{
gaussianCurve(a, b, c, d, xseq)
}
\arguments{
\item{a}{a parameter}

\item{b}{b parameter}

\item{c}{c parameter}

\item{d}{d parameter}

\item{xseq}{the sequence of x values to compute the double gaussian curve for}
}
\description{
Compute the gaussian curve values
}
