% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getABI_Index.R
\name{getABI_Index}
\alias{getABI_Index}
\title{Calculates the index of the ABI fixed grid that corresponds to a geodetic (normal) latitute and longitude}
\usage{
getABI_Index(lat, long, orbitVersion = "NEW")
}
\arguments{
\item{lat}{Latitude in radians (not degrees)}

\item{long}{Longitude in radians (not degrees)}

\item{orbitVersion}{The orbit version ("OLD" if before the satellite moved into its final orbit for GOES-East)}
}
\description{
Calculates the index of the ABI fixed grid that corresponds to a geodetic (normal) latitute and longitude
}
