% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataIndex.R
\name{getDataIndex}
\alias{getDataIndex}
\title{Takes in the y,x coordinates on the ABI grid (vals) and returns a list of the i,j index for the product matrices}
\usage{
getDataIndex(vals, ch, orbitVersion)
}
\arguments{
\item{vals}{The y,x coordinates on the ABI grid}

\item{ch}{The channel number (2, 3, or ACM). Note: NDVI should use channel 3}
}
\description{
Takes in the y,x coordinates on the ABI grid (vals) and returns a list of the i,j index for the product matrices
}
