% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSpecificNDSI.R
\name{getSpecificNDSI}
\alias{getSpecificNDSI}
\title{Function to calculate the NDSI when you have a known pixel}
\usage{
getSpecificNDSI(ind2, ind5, day.time, dataFolder = "GOES_Data2017/")
}
\arguments{
\item{ind2}{The index for the site in the channel 2 grid}

\item{ind5}{The index for the site in the channel 5 grid}

\item{day.time}{The day and time value in the format year,DOY,time (e.g. "20170151200" or "20171821658")}

\item{dataFolder}{The data folder where the data is located (e.g. "GOES_Data2017/")}
}
\description{
Function to calculate the NDSI when you have a known pixel
}
