% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSpecificNDVI.R
\name{getSpecificNDVI}
\alias{getSpecificNDVI}
\title{Function to calculate the NDVI when you have a known pixel}
\usage{
getSpecificNDVI(ind2, ind3, day.time)
}
\arguments{
\item{ind2}{The index for the site in the channel 2 grid}

\item{day.time}{The day and time value in the format year,DOY,time (e.g. "20170151200" or "20171821658")}

\item{ind5}{The index for the site in the channel 5 grid}

\item{dataFolder}{The data folder where the data is located (e.g. "GOES_Data2017/")}
}
\description{
Function to calculate the NDVI when you have a known pixel
}
