% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOES_data.R
\name{GOES_data}
\alias{GOES_data}
\title{For GOES NDVI data, construct the data object for input into MCMC}
\usage{
GOES_data(siteName, lat, long, startDay, endDay, TZ, window = FALSE,
  maxValue = FALSE)
}
\arguments{
\item{siteName}{Site Name}

\item{lat}{Latitude}

\item{long}{Longitude}

\item{startDay}{The start day counted as the day number after 2016-12-31}

\item{endDay}{The end day counted as the day number after 2016-12-31}

\item{TZ}{The timezone off of UTC}

\item{window}{Boolean of if you want to average out over a 4 hour window}
}
\description{
For GOES NDVI data, construct the data object for input into MCMC
}
