% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphMCMC_Outputs.R
\name{graphMCMC_Outputs}
\alias{graphMCMC_Outputs}
\title{Plot the MCMC outputs (rough and not generalized)}
\usage{
graphMCMC_Outputs(outputFileName, siteFileName, iseq, startDay, endDay)
}
\arguments{
\item{outputFileName}{The filename for the graphs}

\item{siteFileName}{The file name of the site data}

\item{iseq}{The row numbers for the desired sites in the siteFileName}

\item{startDay}{The start day}

\item{endDay}{The end day}
}
\description{
Plot the MCMC outputs (rough and not generalized)
}
