% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphMCMC_Outputs_withData.R
\name{graphMCMC_Outputs_withData}
\alias{graphMCMC_Outputs_withData}
\title{Plot the MCMC outputs with data (rough and not generalized)}
\usage{
graphMCMC_Outputs_withData(outputFileName, siteFileName, iseq, startDaySH,
  endDaySH, startDayDB, endDayDB)
}
\arguments{
\item{outputFileName}{The filename for the graphs}

\item{siteFileName}{The file name of the site data}

\item{iseq}{The row numbers for the desired sites in the siteFileName}

\item{startDaySH}{The start day for the shrubland sites}

\item{endDaySH}{The end day for the shrubland sites}

\item{startDayDB}{The start day for the deciduous broadleaf sites}

\item{endDayDB}{The end day for the deciduous broadleaf sites}
}
\description{
Plot the MCMC outputs with data (rough and not generalized)
}
