% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODIS_data.R
\name{MODIS_data}
\alias{MODIS_data}
\title{For MODIS EVI data, construct the data object for input into MCMC}
\usage{
MODIS_data(siteName, lat, long, startDay, endDay, metric)
}
\arguments{
\item{siteName}{Site Name}

\item{lat}{Latitude}

\item{long}{Longitude}

\item{startDay}{The start day counted as the day number after 2016-12-31}

\item{endDay}{The end day counted as the day number after 2016-12-31}

\item{metric}{"NDVI" or "EVI"}
}
\description{
For MODIS EVI data, construct the data object for input into MCMC
}
