% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PC_data.R
\name{PC_data}
\alias{PC_data}
\title{For PhenoCam data, construct the data object for input into MCMC}
\usage{
PC_data(siteName, URL, startDate, endDate, seasonOrder = "AS")
}
\arguments{
\item{siteName}{The site name}

\item{URL}{PhenoCam network URL}

\item{startDate}{The start date}

\item{endDate}{The end date}
}
\description{
For PhenoCam data, construct the data object for input into MCMC
}
