% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitEquations.R
\name{pheno.logistic}
\alias{pheno.logistic}
\title{Compute the logistic curve for phenology values from Zhang et al. (2003)}
\usage{
pheno.logistic(Tran, b, c, d, xseq)
}
\arguments{
\item{Tran}{DOY of transition date}

\item{b}{b parameter (from Zhang et al., 2003)}

\item{c}{c parameter (from Zhang et al., 2003)}

\item{d}{d parameter (from Zhang et al., 2003)}

\item{xseq}{the sequence of x values to compute the logistic curve for}
}
\description{
Compute the logistic curve for phenology values from Zhang et al. (2003)
}
