% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitEquations.R
\name{rescale}
\alias{rescale}
\title{Rescales the values to be between 0 and 1}
\usage{
rescale(c, d, yseq)
}
\arguments{
\item{c}{c parameter (from Zhang et al., 2003)}

\item{d}{d parameter (from Zhang et al., 2003)}

\item{yseq}{the sequence of y values to compute the logistic curve for}
}
\description{
Rescales the values to be between 0 and 1
}
