% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMCMC_Model.R
\name{runMCMC_Model}
\alias{runMCMC_Model}
\title{Continues to run MCMC until convergence and large enough effective sample size}
\usage{
runMCMC_Model(j.model, variableNames, maxIter = 1e+09, baseNum = 80000,
  iterSize = 40000, maxGBR = 10, ID = "", sampleCutoff = 5000)
}
\arguments{
\item{j.model}{The jags model}

\item{variableNames}{A vector of the variable names that need to converge}

\item{maxIter}{The maximum number of iterations to run}

\item{baseNum}{The number of initial iterations to run}

\item{iterSize}{The number of iterations per each run (between when GBR values are checked)}

\item{maxGBR}{The maximum allowable GBR value after the baseNum of iterations are run}

\item{ID}{An optional identifier to print if it doesn't converge}

\item{sampleCutoff}{The minimum desired effective sample size for testing (default is 5000)}
}
\description{
Continues to run MCMC until convergence and large enough effective sample size
}
