% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitEquations.R
\name{shrublandYvals}
\alias{shrublandYvals}
\title{Compute the logistic curve for phenology values from Zhang et al. (2003)}
\usage{
shrublandYvals(Tran, b, c, d, k, r, xseq)
}
\arguments{
\item{Tran}{DOY of transition date}

\item{b}{b parameter from shrubland process model}

\item{c}{c parameter from shrubland process model}

\item{d}{d parameter from shrubland process model}

\item{k}{k parameter from shrubland process model}

\item{r}{r parameter from shrubland process model}

\item{xseq}{the sequence of x values to compute the shrubland curve for}
}
\description{
Compute the logistic curve for phenology values from Zhang et al. (2003)
}
